Revise API to default to user feeds, items, categories and added tests.
