+++ sourcemash/api/feeds.py
@@ -2,12 +2,14 @@ from . import api
 from sourcemash.database import db
 from flask import abort
 from flask.ext.restful import Resource, reqparse, fields, marshal
+from flask.ext.security import current_user, login_required
 
 from datetime import datetime, date
 
 import feedparser
 
 from sourcemash.models import Feed
+from sourcemash.forms import FeedForm
 
 feed_fields = {
     'id': fields.Integer,
@@ -18,8 +20,53 @@ feed_fields = {
 
 class FeedListAPI(Resource):
 
+    def __init__(self):
+        self.reqparse = reqparse.RequestParser()
+        self.reqparse.add_argument('url', type=str, required=True,
+                                    help='No url provided')
+        super(FeedListAPI, self).__init__()
+
+    @login_required
+    def get(self):
+        return {'feeds': [marshal(feed, feed_fields) for feed in current_user.subscribed]}
+
+
+    @login_required
+    def post(self):
+        args = self.reqparse.parse_args()
+        form = FeedForm(obj=args)
+
+        if not form.validate_on_submit():
+            return {"errors": form.errors}, 422
+
+        rss_feed = feedparser.parse(form.url.data)
+
+        try:
+            feed = Feed.query.filter(Feed.url==rss_feed['url']).one()
+        except:
+
+            feed = Feed(title=rss_feed['feed']['title'],
+                        url=rss_feed['url'],
+                        last_updated=datetime.min)
+
+            db.session.add(feed)
+            db.session.commit()
+
+        try:
+            subscription = current_user.subscribed.filter(Feed.id==feed.id).one()
+            return {"errors": {"url": ["Already subscribed"]}}, 409
+        except:
+            current_user.subscribed.append(feed)
+            db.session.commit()
+
+        return marshal(feed, feed_fields), 201
+
+
+class FeedListAllAPI(Resource):
+
     def get(self):
         return {'feeds': [marshal(feed, feed_fields) for feed in Feed.query.all()]}
+
 
 class FeedAPI(Resource):
 
@@ -27,5 +74,21 @@ class FeedAPI(Resource):
         feed = Feed.query.get_or_404(id)
         return {'feed': marshal(feed, feed_fields)}
 
+
+    @login_required
+    def delete(self, id):
+        """Unsubscribe user from feed."""
+
+        try:
+            subscription = current_user.subscribed.filter(Feed.id==id).one()
+        except:
+            abort(404)
+
+        current_user.subscribed.remove(subscription)
+        db.session.commit()
+
+        return {'result': True}
+
 api.add_resource(FeedListAPI, '/feeds', endpoint='feeds')
+api.add_resource(FeedListAllAPI, '/feeds/all', endpoint='feeds_all')
 api.add_resource(FeedAPI, '/feeds/<int:id>', endpoint='feed')
