+++ tests/fixtures/app_fixtures.py
@@ -50,7 +50,7 @@ def session(db, request):
     user_factories.UserFactory._meta.sqlalchemy_session = session
     item_factories.ItemFactory._meta.sqlalchemy_session = session
     item_factories.EbolaItemFactory._meta.sqlalchemy_session = session
-    item_factories.ItemWithCategory._meta.sqlalchemy_session = session
+    item_factories.ItemWithCategoryFactory._meta.sqlalchemy_session = session
     role_factories.RoleFactory._meta.sqlalchemy_session = session
 
     yield db.session
